+++ src/main/java/insane96mcp/survivalreimagined/mixin/BaseSpawnerMixin.java
@@ -1,6 +1,6 @@
 package insane96mcp.survivalreimagined.mixin;
 
-import insane96mcp.survivalreimagined.module.misc.feature.Spawners;
+import insane96mcp.survivalreimagined.module.world.feature.Spawners;
 import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.level.BaseSpawner;