+++ src/main/java/insane96mcp/survivalreimagined/mixin/MonsterMixin.java
@@ -1,7 +1,7 @@
 package insane96mcp.survivalreimagined.mixin;
 
 import insane96mcp.insanelib.base.Feature;
-import insane96mcp.survivalreimagined.module.misc.feature.Spawners;
+import insane96mcp.survivalreimagined.module.world.feature.Spawners;
 import net.minecraft.core.BlockPos;
 import net.minecraft.util.RandomSource;
 import net.minecraft.world.entity.EntityType;