+++ src/main/java/insane96mcp/survivalreimagined/module/combat/feature/ShieldsPlus.java
@@ -6,7 +6,8 @@ import insane96mcp.insanelib.base.Module;
 import insane96mcp.insanelib.base.config.Config;
 import insane96mcp.insanelib.base.config.LoadFeature;
 import insane96mcp.survivalreimagined.module.Modules;
-import insane96mcp.survivalreimagined.setup.IntegratedDataPacks;
+import insane96mcp.survivalreimagined.module.misc.feature.DataPacks;
+import insane96mcp.survivalreimagined.setup.IntegratedDataPack;
 import net.minecraft.server.packs.PackType;
 
 @Label(name = "Shields+", description = "Changes to Shields+.")
@@ -18,6 +19,6 @@ public class ShieldsPlus extends Feature {
 
 	public ShieldsPlus(Module module, boolean enabledByDefault, boolean canBeDisabled) {
 		super(module, enabledByDefault, canBeDisabled);
-		IntegratedDataPacks.INTEGRATED_DATA_PACKS.add(new IntegratedDataPacks.IntegratedDataPack(PackType.SERVER_DATA, "shields", net.minecraft.network.chat.Component.literal("Survival Reimagined Shields"), () -> super.isEnabled() && shieldsPlusCompatDataPack));
+		IntegratedDataPack.INTEGRATED_DATA_PACKS.add(new IntegratedDataPack(PackType.SERVER_DATA, "shields", net.minecraft.network.chat.Component.literal("Survival Reimagined Shields"), () -> super.isEnabled() && !DataPacks.disableAllDataPacks && shieldsPlusCompatDataPack));
 	}
 }